requirements.txt
alembic==1.8.1
backports.zoneinfo==0.2.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
zipp==3.8.1
